added basic crud for assets

+++ app/src/UseCases/Category/DeleteCategoryUseCase.php
@@ -2,6 +2,7 @@
 
 namespace UseCases\Category;
 
+use Repositories\Asset\AssetRepositoryInterface;
 use Repositories\Category\CategoryRepositoryInterface;
 use DomainException;
 use Exception;
@@ -10,8 +11,10 @@ use RuntimeException;
 class DeleteCategoryUseCase
 {
     public function __construct(
-        private CategoryRepositoryInterface $repository
-    ) {}
+        private CategoryRepositoryInterface $repository,
+        private AssetRepositoryInterface $assetRepository
+    ) {
+    }
 
     public function execute(int $id): void
     {
@@ -19,6 +22,10 @@ class DeleteCategoryUseCase
             $category = $this->repository->getById($id);
             if (!$category) {
                 throw new DomainException('Category not found');
+            }
+            $assets = $this->assetRepository->getByCategoryId($id);
+            if ($assets) {
+                throw new DomainException('Category has assets');
             }
             $this->repository->delete($id);
         } catch (RuntimeException $e) {
